cloud_functions/allocation_ratios/requirements.txt
functions-framework==3.*
google-cloud-storage==2.17.0
openpyxl==3.1.3
pandas==2.2.2
pandas-gbq==0.23.1
